+++ src/pyppms/exceptions.py
@@ -1,0 +1,5 @@
+"""PyPPMS exception classes."""
+
+
+class NoDataError(ValueError):
+    """Exception indicating no data was received from PUMAPI."""